@@ -1,11 +1,11 @@
 name: diagonal
 description: Create diagonal clipped widgets easily.
-version: 1.0.1
+version: 2.0.0-nullsafety.0
 author: Figen Güngör <figengungor1@gmail.com>
 homepage: https://github.com/figengungor/diagonal
 
 environment:
-  sdk: '>=1.20.1 <3.0.0'
+  sdk: '>=2.12.0-133.2.beta  <3.0.0'
 
 
 dependencies:
